test: migrate Torch backend tests to separate files for better organization
- Removed the existing `test_torch_backend.py` file and created dedicated test files for CPU, CUDA, MPS, and utility functions.
- This restructuring enhances test organization and maintainability, allowing for clearer separation of concerns across different backend implementations.
- Each new test file includes comprehensive tests for functionality, error handling, and performance validation specific to its respective backend.
